fixedbugs: the lease golden's C side answered unsigned long, which Windows makes 32 bits, and the boundary refused it
Found on the Windows box before the push, by the instrument built for it:
tests/golden/run/lease-c-keeps-the-pointer.h declared stash_len as returning
unsigned long, the .hero declared u64, and the FFI's own _Static_assert refused
the pair on Windows -- error[ffi_return_type]: stash_len does not return u64,
that is what the header says, and clang read it. On the Mac and on Linux
unsigned long is 64 bits and the pair agreed. size_t is what strlen answers on
all three, so the header says size_t. A platform fact is run on a platform or it
is an inference (.claude/rules/platforms.md), and this one was run: 13 13 13 on
the box after the fix, 618 own tests and 125 harness tests there, and the abort
case exits 127 on Windows exactly as the shipped abort golden does.

+++ tests/golden/run/lease-c-keeps-the-pointer.h
@@ -4,4 +4,4 @@
 #include <string.h>
 static const char *held_by_c;
 static void stash_put(const char *s) { held_by_c = s; }
-static unsigned long stash_len(void) { return strlen(held_by_c); }
+static size_t stash_len(void) { return strlen(held_by_c); }
